microservice-userauth/requirements.txt
Werkzeug==3.0.1
grpcio==1.62.2
grpcio-tools==1.62.2
protobuf==4.25.3